# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application (PPA V5) API
  version: 1.0.0
  description: |
    RESTful API for a campus placement management system.
    Supports three roles: Admin, Company, and Student.
servers:
  - url: http://localhost:5001
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        full_name:
          type: string
        department:
          type: string
        cgpa:
          type: number
        graduation_year:
          type: integer
        resume_path:
          type: string
        phone:
          type: string
        bio:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        is_blacklisted:
          type: boolean
        skills:
          type: array
          items:
            type: string
    CompanyProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        company_name:
          type: string
        hr_name:
          type: string
        hr_email:
          type: string
        hr_phone:
          type: string
        website:
          type: string
        description:
          type: string
        industry:
          type: string
        company_size:
          type: string
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        email:
          type: string
        is_active:
          type: boolean
        is_blacklisted:
          type: boolean
    PlacementDrive:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        company_name:
          type: string
        drive_name:
          type: string
        job_title:
          type: string
        job_description:
          type: string
        eligibility_branch:
          type: string
        min_cgpa:
          type: number
        eligible_year:
          type: integer
        application_deadline:
          type: string
          format: date-time
        location:
          type: string
        salary:
          type: string
        job_type:
          type: string
          enum: [Full-time, Internship, Contract]
        status:
          type: string
          enum: [pending, approved, rejected, closed]
        created_at:
          type: string
          format: date-time
        application_count:
          type: integer
    Application:
      type: object
      properties:
        id:
          type: integer
        student_id:
          type: integer
        drive_id:
          type: integer
        student_name:
          type: string
        student_email:
          type: string
        student_cgpa:
          type: number
        student_department:
          type: string
        drive_name:
          type: string
        company_name:
          type: string
        job_title:
          type: string
        application_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected, withdrawn]
        interview_type:
          type: string
        remarks:
          type: string
        cover_letter:
          type: string
...
    MonthlyReport:
      type: object
      properties:
        id:
          type: integer
        month:
          type: string
        total_drives:
          type: integer
        total_applications:
          type: integer
        total_selected:
          type: integer
        report_path:
          type: string
        created_at:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        message:
          type: string
paths:
...
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Get admin dashboard statistics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard stats with chart data
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_students:
                    type: integer
                  total_companies:
                    type: integer
                  total_drives:
                    type: integer
                  total_applications:
                    type: integer
                  selected_students:
                    type: integer
                  pending_companies:
                    type: integer
                  pending_drives:
                    type: integer
                  blacklisted_users:
                    type: integer
                  chart_app_status:
                    type: object
                  chart_drive_status:
                    type: object
  /api/admin/search:
    get:
      tags: [Admin]
      summary: Search companies or students
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [companies, students]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
  /api/admin/companies:
    get:
      tags: [Admin]
      summary: List all companies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of company profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyProfile"
  /api/admin/companies/pending:
    get:
      tags: [Admin]
      summary: List pending companies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Companies awaiting approval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyProfile"
  /api/admin/companies/{id}/approve:
    put:
      tags: [Admin]
      summary: Approve a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/companies/{id}/reject:
    put:
      tags: [Admin]
      summary: Reject a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks:
                  type: string
      responses:
        "200":
          description: Company rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/companies/{id}/blacklist:
    put:
      tags: [Admin]
      summary: Blacklist or unblacklist a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [blacklist, unblacklist]
      responses:
        "200":
          description: Blacklist status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/drives:
    get:
      tags: [Admin]
      summary: List all placement drives
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, closed]
      responses:
        "200":
          description: List of drives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacementDrive"
  /api/admin/drives/pending:
    get:
      tags: [Admin]
      summary: List pending drives
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Drives awaiting approval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacementDrive"
  /api/admin/drives/{id}/approve:
    put:
      tags: [Admin]
      summary: Approve a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drive approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/drives/{id}/reject:
    put:
      tags: [Admin]
      summary: Reject a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks:
                  type: string
      responses:
        "200":
          description: Drive rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/drives/{id}/close:
    put:
      tags: [Admin]
      summary: Close a drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drive closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/students:
    get:
      tags: [Admin]
      summary: List all students
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of student profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
  /api/admin/students/{id}/deactivate:
    put:
      tags: [Admin]
      summary: Activate or deactivate a student
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [activate, deactivate]
      responses:
        "200":
          description: Student status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/students/{id}/blacklist:
    put:
      tags: [Admin]
      summary: Blacklist or unblacklist a student
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [blacklist, unblacklist]
      responses:
        "200":
          description: Blacklist status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/admin/applications:
    get:
      tags: [Admin]
      summary: List all applications (with optional filters)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [applied, shortlisted, selected, rejected, withdrawn]
        - name: drive_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /api/admin/reports/monthly:
    get:
      tags: [Admin]
      summary: List monthly reports
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"
  /api/admin/reports/generate:
    post:
      tags: [Admin]
      summary: Trigger monthly report generation (async Celery task)
      security:
        - BearerAuth: []
      responses:
        "202":
          description: Report generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job_id:
                    type: integer
  /api/admin/reports/download/{id}:
    get:
      tags: [Admin]
      summary: Download a report PDF
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ─── Company ──────────────────────────────────────────
...